+++ packages/tupaia-web/src/utils/index.ts
@@ -10,5 +10,4 @@ export { useEntityLink } from './useEntityLink';
 export { useDateRanges } from './useDateRanges';
 export { gaEvent } from './ga';
 export { transformDownloadLink } from './transformDownloadLink';
-export { useHandleNoAccessError } from './useHandleNoAccessError';
 export { useDebounce } from './useDebounce';
