+++ packages/ui-components/src/components/Loaders/FullPageLoader.tsx
@@ -5,7 +5,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import CircularProgress from '@material-ui/core/CircularProgress';
-import { FlexCenter } from './Layout';
+import { FlexCenter } from '../Layout';
 
 const LoadingContainer = styled(FlexCenter)`
   height: 100vh;
