+++ packages/ui-components/src/components/Loaders/SpinningLoader.tsx
@@ -5,7 +5,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import CircularProgress from '@material-ui/core/CircularProgress';
-import { FlexCenter } from './Layout';
+import { FlexCenter } from '../Layout';
 
 const Container = styled(FlexCenter)`
   width: 100%;
